Concat all into one, until TS concat supports commonjs
